GRID-2312 Sh:defaultValue
Add Parameter support for an optional `defaultValue`.
`TermParameter` uses generic type for `defaultvalue` and defaults to
`null` itself. This way we only have to replace the return value of
`handleInner` to get it working for optional param values.

All specific implementations use overloading on the constructor as this
`defaultValue` is optional